+++ app/src/main/java/org/breezyweather/common/source/ConfigurableSource.kt
@@ -20,7 +20,8 @@ import android.content.Context
 import org.breezyweather.common.preference.Preference
 
 /**
- * Implement this if you need a preference screen
+ * Implement this if you need a preference screen for all locations
+ * Use PreferencesParametersSource instead if you need per-location parameters
  */
 interface ConfigurableSource : Source {