+++ app/src/main/java/org/breezyweather/sources/openmeteo/json/OpenMeteoWeatherResult.kt
@@ -27,5 +27,7 @@ data class OpenMeteoWeatherResult(
     @SerialName("current") val current: OpenMeteoWeatherCurrent? = null,
     val daily: OpenMeteoWeatherDaily? = null,
     val hourly: OpenMeteoWeatherHourly? = null,
-    @SerialName("minutely_15") val minutelyFifteen: OpenMeteoWeatherMinutely? = null
+    @SerialName("minutely_15") val minutelyFifteen: OpenMeteoWeatherMinutely? = null,
+    val error: Boolean? = null,
+    val reason: String? = null
 )