+++ app/src/src_nofdroid/org/breezyweather/sources/metie/MetIeService.kt
@@ -25,7 +25,7 @@ import breezyweather.domain.weather.wrappers.WeatherWrapper
 import org.breezyweather.common.exceptions.InvalidLocationException
 import org.breezyweather.common.source.HttpSource
 import org.breezyweather.common.source.MainWeatherSource
-import org.breezyweather.common.source.ParameterizedLocationSource
+import org.breezyweather.common.source.LocationParametersSource
 import org.breezyweather.common.source.ReverseGeocodingSource
 import org.breezyweather.common.source.SecondaryWeatherSource
 import org.breezyweather.common.source.SecondaryWeatherSourceFeature
@@ -40,7 +40,7 @@ import javax.inject.Inject
 class MetIeService @Inject constructor(
     client: Retrofit.Builder
 ) : HttpSource(), MainWeatherSource, SecondaryWeatherSource,
-    ReverseGeocodingSource, ParameterizedLocationSource {
+    ReverseGeocodingSource, LocationParametersSource {
 
     override val id = "metie"
     override val name = "MET Éireann"